Fix to segfault that occurred when importing rxd with no access to the window system. This was a problem in ssh without X11 tunneling and with MPI.
The basic problem was that the GUI attempts to pre-prepare certain dialogs, but that
means that it uses NEURON's GUI tools which do not always fail silently when no GUI
is available.

The current solution is to delay importing the GUI until it is requested, thereby
delaying the initial creation of the dialogs.